Svelto Integrated: Success!
The schedulers are now all setup, the composition root is setup, a test entity is being created properly, the rendering engine works correctly.

The next step is to move the entire engine over to the entity system, get a scene system in the works, and all of that fun stuff!