Material system, WIP new Renderer, Code Cleanup
Material System : This system is designed to be a wrapper for shaders and access to their properties. They list the capabilities they have, and the renderer will be able to provide them with the information they desire.

WIP New Renderer : The new renderer, which is going to allow rendering "NOW" and when the renderer normally fires, is going to support the material system, is going to allow transparent objects to be rendered correctly (look into Octree), allow multiple cameras, allow the core renderer to use orthographic cameras, multiple lights, and probably more. After some discussion with some people on the C# Discord, I have been told I should be rendering based on an Octree, probably rendering furthest to closest, possibly have transparent be a second pass if needed.

Code cleanup : I ran the code cleaner tool over the codebase to get it cleaned up a bit, and changed things like how the camera and light properties are accessed, we aren't working in Java you know!